#334 Refactor geometry into separate directory
