+++ nexus_constructor/geometry/cylindrical_geometry.py
@@ -13,102 +13,9 @@ from nexus_constructor.ui_utils import (
     numpy_array_to_qvector3d,
     qvector3d_to_numpy_array,
 )
-from typing import List, TypeVar, Tuple
-
-
-# Temporary method here until the one above is no longer needed
-def validate_nonzero_qvector(value: QVector3D):
-    if value.x() == 0 and value.y() == 0 and value.z() == 0:
-        raise ValueError("Vector is zero length")
-
-
-def _get_an_orthogonal_unit_vector(input_vector: QVector3D):
-    """
-    Return a unit vector which is orthogonal to the input vector
-    There are infinite valid solutions, just one is returned
-    """
-    if np.abs(input_vector.z()) < np.abs(input_vector.x()):
-        vector = QVector3D(input_vector.x(), -input_vector.y(), 0.0)
-        return vector.normalized()
-    vector = QVector3D(0.0, -input_vector.z(), input_vector.y())
-    return vector.normalized()
-
-
-class OFFGeometry:
-    """
-    Stores arbitrary 3D geometry as a list of vertices and faces, based on the Object File Format
-
-    vertices:   list of Vector objects used as corners of polygons in the geometry
-    faces:  list of integer lists. Each sublist is a winding path around the corners of a polygon. Each sublist item is
-            an index into the vertices list to identify a specific point in 3D space
-    """
-
-    geometry_str = "OFF"
-
-    def __init__(
-        self,
-        vertices: List[QVector3D] = None,
-        faces: List[List[int]] = None,
-        units: str = "",
-        file_path: str = "",
-    ):
-        self.vertices = vertices
-        self.faces = faces
-
-        # Source units and file path are retained only for the purpose of populating the edit component window
-        # with the options previously chosen by the user
-        self.units = units
-        self.file_path = file_path
-
-    @property
-    def winding_order(self):
-        return [point for face in self.faces for point in face]
-
-    @property
-    def winding_order_indices(self):
-        face_sizes = [len(face) for face in self.faces]
-        return [sum(face_sizes[0:i]) for i in range(len(face_sizes))]
-
-    @property
-    def off_geometry(self):
-        return self
-
-
-OFFCube = OFFGeometry(
-    vertices=[
-        QVector3D(-0.5, -0.5, 0.5),
-        QVector3D(0.5, -0.5, 0.5),
-        QVector3D(-0.5, 0.5, 0.5),
-        QVector3D(0.5, 0.5, 0.5),
-        QVector3D(-0.5, 0.5, -0.5),
-        QVector3D(0.5, 0.5, -0.5),
-        QVector3D(-0.5, -0.5, -0.5),
-        QVector3D(0.5, -0.5, -0.5),
-    ],
-    faces=[
-        [0, 1, 3, 2],
-        [2, 3, 5, 4],
-        [4, 5, 7, 6],
-        [6, 7, 1, 0],
-        [1, 7, 5, 3],
-        [6, 0, 2, 4],
-    ],
-)
-
-
-class NoShapeGeometry:
-    """
-    Dummy object for components with no geometry.
-    """
-
-    geometry_str = "None"
-
-    def __init__(self):
-        pass
-
-    @property
-    def off_geometry(self):
-        return OFFCube
+from nexus_constructor.geometry.utils import validate_nonzero_qvector
+from nexus_constructor.geometry.off_geometry import OFFGeometry
+from typing import Tuple
 
 
 class CylindricalGeometry:
