+++ nexus_constructor/off_renderer.py
@@ -7,7 +7,7 @@ https://github.com/geehalel/npindi/blob/57c092200dd9cb259ac1c730a1258a378a1a6342
 """
 
 from nexus_constructor.pixel_data import PixelData, PixelGrid
-from nexus_constructor.geometry_types import OFFGeometry
+from nexus_constructor.geometry import OFFGeometry
 from PySide2.Qt3DRender import Qt3DRender
 from PySide2.QtGui import QVector3D
 import struct
