+++ tests/test_geometry_loader.py
@@ -1,5 +1,5 @@
 from nexus_constructor.pixel_data import PixelGrid
-from nexus_constructor.geometry_types import OFFGeometry
+from nexus_constructor.geometry import OFFGeometry
 from nexus_constructor.geometry_loader import load_geometry_from_file_object
 from nexus_constructor.off_renderer import QtOFFGeometry
 from PySide2.QtGui import QVector3D
